tests/wpt/tests/html/editing/dnd/interactiveelements/input_button_inside_draggable_element.html

<!doctype html>
<html>
  <head>
    <title>Input button inside draggable element</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/resources/testdriver.js"></script>
    <script src="/resources/testdriver-vendor.js"></script>
    <script src="/resources/testdriver-actions.js"></script>
    <style type="text/css">
#inner { border: 1px solid orange; border-top-width: 20px; }
    </style>
  </head>
  <body>

    <p>Press your mouse button down on the orange block and drag downwards. It should drag the draggable element.</p>
    <p>Press your mouse button down on the text and drag rightward. It should drag the draggable element.</p>
    <div draggable="true" id="outer">
      <input type="button" value="Dummy text" id="inner">
    </div>

    <script>
    const outer = document.getElementById("outer");
    const inner = document.getElementById("inner");

    function testDragInputButton(dragActionFun, msg) {
      promise_test(t => {
        return new Promise((resolve, reject) => {
          let result = {};

          const dragStartListener = ev => {
            ev.preventDefault();
            result.didReceiveDragStart = true;
            result.targetId = ev.target.id;
            result.dataTransferItemsCount = ev.dataTransfer.items.length;
            result.dataTransferTypesCount = ev.dataTransfer.types.length;
          };
          outer.addEventListener("dragstart", dragStartListener, { once: true });
          t.add_cleanup(() => outer.removeEventListener("dragstart", dragStartListener));

          outer.addEventListener("mouseup", () => {
            resolve(result);
          }, { once: true });

          dragActionFun();
        }).then(result => {
          assert_true(result.didReceiveDragStart, "dragstart should not fire");
          assert_equals(result.targetId, "outer", "should drag outer element");
          assert_equals(result.dataTransferItemsCount, 0, "dataTransfer should have no items");
          assert_equals(result.dataTransferTypesCount, 0, "dataTransfer should have no types");
        });
      }, msg);
    }

    testDragInputButton(async () => {
      const centerH = inner.getBoundingClientRect().height / 2;
      new test_driver.Actions()
        // Move pointer to the center of the top-border.
        .pointerMove(0, 10 - centerH, { origin: inner })
        .pointerDown()
        // Move pointer to the center of the button.
        .pointerMove(0, 0, { origin: inner })
        .pointerUp()
        .send();
    }, "Input button inside draggable element should be draggable from border");

    testDragInputButton(async () => {
      const centerW = inner.getBoundingClientRect().width / 2;
      new test_driver.Actions()
        // Move pointer to the start of the button text.
        .pointerMove(5 - centerW, 10, { origin: inner })
        .pointerDown()
        // Move pointer rightward.
        .pointerMove(0, 10, { origin: inner })
        .pointerUp()
        .send();
    }, "Input button inside draggable element should be draggable from button text");
    </script>

  </body>
</html>
